+++ src/compiler/emitter/mod.rs
@@ -10,5 +10,7 @@ mod lines_2500_3000;
 pub use lines_2500_3000::*;
 mod lines_3000_3500;
 pub use lines_3000_3500::*;
-mod lines_3500_end;
-pub use lines_3500_end::*;
+mod lines_3500_4000;
+pub use lines_3500_4000::*;
+mod lines_4000_end;
+pub use lines_4000_end::*;
